# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{slug}": {
      "get": {
        "operationId": "ProjectsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a project by slug",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Project not found"
          },
          "409": {
            "description": "Conflict - duplicate slug or key"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a project (admin only)",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project soft deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete a project (admin only)",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{slug}/members": {
      "get": {
        "operationId": "ProjectMembersController_list",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "current",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of members: { total, current, size, hasNext, hasPrev, records, canManage }"
          },
          "403": {
            "description": "Not a project member"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List project members (any member)",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "ProjectMembersController_add",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberDto"
                }
              }
            }
          },
          "403": {
            "description": "Project admin role required"
          },
          "404": {
            "description": "Project or user not found"
          },
          "409": {
            "description": "Already a member"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an existing user by email (project admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{slug}/members/{userId}": {
      "patch": {
        "operationId": "ProjectMembersController_updateRole",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberDto"
                }
              }
            }
          },
          "404": {
            "description": "Not a member"
          },
          "409": {
            "description": "Would leave the project without an ADMIN"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a member's project role (project admin)",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectMembersController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed"
          },
          "404": {
            "description": "Not a member"
          },
          "409": {
            "description": "Would leave the project without an ADMIN"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a member (project admin)",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Koda API",
    "description": "Dev ticket tracker API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "description": "Project name",
            "example": "Updated Project Name"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "description": "Project slug - lowercase alphanumeric with hyphens",
            "example": "updated-project"
          },
          "key": {
            "type": "string",
            "pattern": "^[A-Z]{2,6}$",
            "description": "Project key - 2-6 uppercase letters",
            "example": "UPDATED"
          },
          "description": {
            "type": "string",
            "description": "Project description",
            "example": "Updated description"
          },
          "gitRemoteUrl": {
            "type": "string",
            "format": "uri",
            "description": "Git remote repository URL",
            "example": "https://github.com/nathapp-io/updated-repo"
          },
          "autoIndexOnClose": {
            "type": "boolean",
            "description": "Automatically index commits on ticket close",
            "example": false
          },
          "autoAssign": {
            "type": "string",
            "enum": [
              "OFF",
              "SUGGEST",
              "AUTO"
            ],
            "description": "Auto-assign mode for tickets: OFF, SUGGEST, or AUTO",
            "example": "SUGGEST"
          },
          "ciWebhookToken": {
            "type": "string",
            "minLength": 32,
            "description": "CI webhook token for authentication",
            "example": "ci_webhook_token_123"
          },
          "graphifyEnabled": {
            "type": "boolean",
            "description": "Enable graphify feature for this project",
            "example": false
          }
        }
      },
...
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email of an existing user",
            "example": "dev@example.com"
          },
          "role": {
            "enum": [
              "ADMIN",
              "DEVELOPER",
              "VIEWER"
            ],
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "ProjectMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "DEVELOPER",
              "AGENT",
              "VIEWER"
            ]
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "email",
          "role",
          "joinedAt"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "ADMIN",
              "DEVELOPER",
              "VIEWER"
            ],
            "type": "string"
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}